+++ litho-core/src/main/java/com/facebook/litho/ComponentHost.java
@@ -799,6 +799,8 @@ public class ComponentHost extends Host implements DisappearingHost {
         ((Drawable) content).draw(canvas);
       }
     }
+
+    DebugDraw.draw(this, canvas);
   }
 
   @Override